feat(core): Update Resource Mapping delete to use get before delete for cli output (#398)
Due to changes made in https://github.com/opentdf/platform/pull/1422 to
only return an object with the ID for delete operations, we need to
update the otdfctl Resource Mapping delete operation to make a Get
request from the DB before deletion of the record. The object returned
from the Get is then used for the user-friendly CLI output of the whole
object.